+++ src/commonMain/kotlin/at/posselt/pfrpg2e/kingdom/modifiers/bonuses/RepairBonuses.kt
@@ -2,7 +2,7 @@ package at.posselt.pfrpg2e.kingdom.modifiers.bonuses
 
 import at.posselt.pfrpg2e.kingdom.modifiers.Modifier
 import at.posselt.pfrpg2e.kingdom.modifiers.ModifierType
-import at.posselt.pfrpg2e.kingdom.modifiers.expressions.HasFlag
+import at.posselt.pfrpg2e.kingdom.modifiers.expressions.HasRollOption
 
 fun createRepairBonus() = Modifier(
     id = "repair-structure-item-bonus",
@@ -11,6 +11,6 @@ fun createRepairBonus() = Modifier(
     enabled = true,
     value = 2,
     applyIf = listOf(
-        HasFlag("repair-structure")
+        HasRollOption("repair-structure")
     )
 )