+++ src/jsMain/kotlin/at/posselt/pfrpg2e/migrations/Migrations.kt
@@ -21,6 +21,7 @@ import at.posselt.pfrpg2e.migrations.migrations.Migration12
 import at.posselt.pfrpg2e.migrations.migrations.Migration13
 import at.posselt.pfrpg2e.migrations.migrations.Migration14
 import at.posselt.pfrpg2e.migrations.migrations.Migration15
+import at.posselt.pfrpg2e.migrations.migrations.Migration16
 import at.posselt.pfrpg2e.migrations.migrations.Migration9
 import at.posselt.pfrpg2e.settings.pfrpg2eKingdomCampingWeather
 import at.posselt.pfrpg2e.utils.getAppFlag
@@ -65,6 +66,7 @@ private val migrations = listOf(
     Migration13(),
     Migration14(),
     Migration15(),
+    Migration16(),
 )
 
 val latestMigrationVersion = migrations.maxOfOrNull { it.version }!!